requirements.txt
# Data manipulation
pandas==2.1.4
numpy==1.26.3

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.16.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Visualization
streamlit==1.30.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Data transformation
dbt-postgres==1.7.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
